Move `Version` and `YamlConfigFile` to dbt_semantic_interfaces.parsing
The `objects` dir of `dbt_semantic_interfaces` is becoming the `implementations`
dir. What that means is that things in `objects` should only be things that
exist on what is currently called the `UserConfiguredModel`, soon to be called
the `SemanticManifest`. These objects, `Version` and `YamlConfigFile` are only
used during parsing, and are not part of the implemented `UserConfiguredModel`,
thus we moved them.
