+++ metricflow/test/model/validations/helpers.py
@@ -2,12 +2,13 @@ import textwrap
 from typing import Optional, Sequence
 
 from metricflow.engine.models import Dimension
-from dbt_semantic_interfaces.objects.common import FileSlice, Metadata, YamlConfigFile
+from dbt_semantic_interfaces.objects.common import FileSlice, Metadata
 from dbt_semantic_interfaces.objects.constraints.where import WhereClauseConstraint
 from dbt_semantic_interfaces.objects.data_source import DataSource, DataSourceOrigin, Mutability
 from dbt_semantic_interfaces.objects.elements.entity import Entity
 from dbt_semantic_interfaces.objects.elements.measure import Measure
 from dbt_semantic_interfaces.objects.metric import Metric, MetricType, MetricTypeParams
+from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
 
 
 def base_model_file() -> YamlConfigFile:
