+++ metricflow/test/model/validations/test_entities.py
@@ -6,11 +6,11 @@ from typing import Callable
 import pytest
 
 from metricflow.model.model_validator import ModelValidator
-from dbt_semantic_interfaces.objects.common import YamlConfigFile
 from dbt_semantic_interfaces.objects.data_source import DataSource
 from dbt_semantic_interfaces.objects.elements.entity import EntityType
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.parsing.dir_to_model import parse_yaml_files_to_validation_ready_model
+from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
 from metricflow.model.validations.entities import (
     NaturalEntityConfigurationRule,
     OnePrimaryEntityPerDataSourceRule,
