+++ metricflow/test/query/test_query_parser_suggestions.py
@@ -6,8 +6,8 @@ import textwrap
 
 import pytest
 
+from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
 from metricflow.errors.errors import UnableToSatisfyQueryError
-from dbt_semantic_interfaces.objects.common import YamlConfigFile
 from metricflow.plan_conversion.time_spine import TimeSpineSource
 from metricflow.test.fixtures.model_fixtures import query_parser_from_yaml
